+++ examples/05_device_functions.cpp
@@ -12,9 +12,9 @@ void axpy(int N, T alpha, const T* x, T* y) {
 
 int main() {
   const int vector_length = 2000;
-  std::vector<double> x_host(vector_length, 1.0);
-  std::vector<double> y_host(vector_length, 0.0);
-  const double alpha = 1.0;
+  std::vector<float> x_host(vector_length, 1.0);
+  std::vector<float> y_host(vector_length, 0.0);
+  const float alpha = 1.0;
 
   // Here we are calling the function from the host,
   // so it has been compiled as regular C++ function
@@ -24,10 +24,10 @@ int main() {
   sycl::context sycl_context{sycl_device};
   sycl::queue sycl_queue{sycl_context, sycl_device};
 
-  double* x =
-      sycl::malloc_device<double>(vector_length, sycl_device, sycl_context);
-  double* y =
-      sycl::malloc_device<double>(vector_length, sycl_device, sycl_context);
+  float* x =
+      sycl::malloc_device<float>(vector_length, sycl_device, sycl_context);
+  float* y =
+      sycl::malloc_device<float>(vector_length, sycl_device, sycl_context);
 
   sycl::event copy_x = sycl_queue.copy(x_host.data(), x, x_host.size());
   sycl::event copy_y = sycl_queue.copy(y_host.data(), y, y_host.size());
@@ -41,8 +41,8 @@ int main() {
     cgh.parallel_for({vector_length / thread_vector_length},
                      [=](sycl::item<1> work_item) {
                        // These are private to each work-item
-                       double x_thread[thread_vector_length];
-                       double y_thread[thread_vector_length];
+                       float x_thread[thread_vector_length];
+                       float y_thread[thread_vector_length];
 
                        int i = work_item.get_linear_id();
                        int r = work_item.get_range(0);
