+++ examples/07_local_memory.cpp
@@ -11,9 +11,9 @@ int main() {
   constexpr int K{64};
 
   // Linear arrays to store matrices
-  std::vector<double> A_host(M * K, 1.0);
-  std::vector<double> B_host(K * N, 1.0);
-  std::vector<double> C_host(M * N);
+  std::vector<float> A_host(M * K, 1.0);
+  std::vector<float> B_host(K * N, 1.0);
+  std::vector<float> C_host(M * N);
 
   // Create a sycl::queue using the default device selector
   sycl::device sycl_device{sycl::default_selector()};
@@ -21,12 +21,12 @@ int main() {
   sycl::queue sycl_queue{sycl_context, sycl_device};
 
   // Allocate matrices on the device
-  double* A =
-      sycl::malloc_device<double>(A_host.size(), sycl_device, sycl_context);
-  double* B =
-      sycl::malloc_device<double>(B_host.size(), sycl_device, sycl_context);
-  double* C =
-      sycl::malloc_device<double>(C_host.size(), sycl_device, sycl_context);
+  float* A =
+      sycl::malloc_device<float>(A_host.size(), sycl_device, sycl_context);
+  float* B =
+      sycl::malloc_device<float>(B_host.size(), sycl_device, sycl_context);
+  float* C =
+      sycl::malloc_device<float>(C_host.size(), sycl_device, sycl_context);
 
   // Copy from the host to the device
   sycl::event copy_a = sycl_queue.copy(A_host.data(), A, A_host.size());
@@ -55,7 +55,7 @@ int main() {
       auto work_group = work_item.get_group();
 
       // Allocate SLM to use as an explicit cache
-      using tile_t = double[tile_size][block_size];
+      using tile_t = float[tile_size][block_size];
       tile_t& A_tile =
           *dpcpp::group_local_memory_for_overwrite<tile_t>(work_group);
       tile_t& B_tile =
@@ -64,7 +64,7 @@ int main() {
       // Compute C = A * B
       // Each work-group will compute a 16x16 block of C
       // Tile the k-loop by a factor of 8
-      double C_ij{};
+      float C_ij{};
       for (int k_tile{}; k_tile < K; k_tile += tile_size) {
         // Here j plays the role of k
         if (j < tile_size) {
@@ -101,7 +101,7 @@ int main() {
 
   // Verify the results
   for (const auto& C_ij : C_host) {
-    if (static_cast<double>(K) != C_ij) {
+    if (static_cast<float>(K) != C_ij) {
       std::cout << "Verification failed!\n";
       return EXIT_FAILURE;
     }
